Completed plugin configuration to invoke Lambda function after project deploys and after file is uploaded.
