+++ handler.js
@@ -6,9 +6,8 @@ const s3 = new AWS.S3();
 const ddb = new AWS.DynamoDB.DocumentClient();
 
 module.exports.updateDynamoDb = (event, context, callback) => {
-  console.log('Ok');
-  const bucketName = 'cereal-bucket'; // Name of Bucket
-  const keyName = 'shortCereals.json'; // File Name in Bucket
+  const bucketName = 'cereal-bucket';
+  const keyName = 'shortCereals.json';
   const params = { Bucket: bucketName, Key: keyName };
 
   s3.getObject(params, (err, data) => {
