+++ serverless.yml
@@ -42,6 +42,7 @@ provider:
 
 plugins:
   - serverless-s3-sync
+  - invoke-lambda
 
 custom:
   s3Sync:
@@ -82,4 +83,5 @@ package:
     - node_modules
     - package-lock.json
     - package.json
-    - upload_files
+    - upload_files/**
+    - .serverless_plugins
